Tue, Oct 24, 2023, 5:29 AM +02:00

+++ recover/recover.c
@@ -65,7 +65,7 @@ int main(int argc, char *argv[])
             {
                 continue;
             }
-            int* outfile_ptr[counter_image] = (int*) malloc(memo_block * sizeof(BLOCK_SIZE));
+            int* outfile_ptr[counter_image] = (BYTE*) malloc(memo_block * sizeof(BLOCK_SIZE));
             fwrite(buffer, 1, BLOCK_SIZE, outfile);
             memo_block++;
         }
